requirements.txt
Cython==3.0.8
matplotlib==3.8.0
numpy==1.26.4
Pillow==10.3.0
psutil==5.9.0
PyAutoGUI==0.9.54
PyOpenGL==3.1.7
pyopenvdb==0.1.4
PyQt5==5.15.10
PyQt5_sip==12.13.0
pyrr==0.10.3
scipy==1.11.4
skimage
wxPython==4.2.1
